refactor: Change folder names
