+++ renderer/src/components/User/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import useUser from '@hooks/Users/useUser';
-import Skeleton from '@components/Common/Skeleton';
+import useUser from '@hooks/user/useUser';
+import Skeleton from '@components/common/Skeleton';
 import UserCard from './UserCard';
 import UserDetail from './UserDetail';
 import useActiveSection from '@hooks/common/useActiveSection';
-import userLanguageList from '@hooks/languages/useLanguageList';
+import userLanguageList from '@hooks/language/useLanguageList';
 
 function UserSection() {
   const { data: user } = useUser('youthfulhps');
